backend/requirements.txt
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
pytz==2018.5
